requirements.txt
Flask==2.3.2
gunicorn==21.2.0
mysql-connector-python==8.1.0
requests==2.32.0


